chore: rename secondary header file
